Task 12/requirements.txt
pandas
sentence-transformers
faiss-cpu
numpy
flask
